Redesign: algorithms.js
* Renamed PredictionService to RecommendationService.
* Split functionality of algorithms into SimilarityService and RecommendationService.
